pyproject.toml
[project]
name = "weather-app"
version = "0.5.10"
description = "A CLI weather app using OpenWeatherMap API, written in Python 3"
authors = [
    {name = "Lord Sohron",email = "29570082+vihoma@users.noreply.github.com"}
]
license = {text = "BSD 3-Clause"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyowm (>=3.4.0,<4.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "cachetools (>=6.2.0,<7.0.0)",
    "geopy (>=2.4.0,<3.0.0)",
    "setuptools (>=80.9.0, <81.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "async-timeout (>=5.0.0,<6.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "click (>=8.1.0,<9.0.0)"
]

[project.scripts]
# This is the main script entry point for the weather app
weather = "weather_app.main:main"

[tool.poetry]
packages = [
    {include = "weather_app", from = "src"}
]

[tool.poetry.group.dev.dependencies]
pylint = "^4.0.2"
isort = "^7.0.0"
flake8 = "^7.3.0"
ruff = "^0.14.1"
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-mock = "^3.15.1"
mypy = "^1.18.2"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short"
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, test interactions)",
    "functional: Functional tests (end-to-end)",
    "slow: Slow running tests",
    "api: Tests that require external API calls",
    "async: Async tests",
    "skip: Tests that should be skipped"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
